refactor(convex): Move run create and tool-job begin into lib
Keep the agentRuntime:* Convex paths, but pull queued-run insert,
failed-start matching, and tool-job begin out of the wrapper file.
Prompt lookup lives with the transcript key helpers.
